+++ docs/release_notes.md
@@ -1,5 +1,10 @@
 ## DigiNode Tools Release Notes
 
+DigiNode Tools v0.9.12 - 2024-04-10
+- Fix: Hide displayed download progress when checking for DigiByte Core update at Dashboard launch
+- New: Updated Raspberry Pi detection to identify newer Pi 4, Pi CM4, Pi 400 and Pi 5 models.
+- New: DigiNode Tools must be running the latest version to install the other updates. This helps to avoid any install changes between versions.
+
 DigiNode Tools v0.9.11 - 2024-04-02
 - New: Add support for DigiByte Core v8.22.0-rc4